Extracting Vocabulary::Item to module
I prefer to have the various Item types at the same level. This
refactoring accounts for that desire.
