+++ lib/locabulary/items/administrative_unit.rb
@@ -1,10 +1,10 @@
-require 'locabulary/item'
+require 'locabulary/items/base'
 module Locabulary
   module Items
     # Responsible for exposing the data structure logic of the Administrative Units
     #
     # @see ./data/administrative_units.json
-    class AdministrativeUnit < Locabulary::Item
+    class AdministrativeUnit < Locabulary::Items::Base
       configure do |config|
         config.attribute_names = [
           :predicate_name, :term_label, :term_uri, :description, :grouping, :affiliation, :default_presentation_sequence,
