+++ lib/locabulary/json_creator.rb
@@ -2,7 +2,7 @@ require "google/api_client"
 require "google_drive"
 require 'highline/import'
 require 'locabulary'
-require 'locabulary/item'
+require 'locabulary/items'
 require 'json'
 
 module Locabulary
@@ -61,7 +61,7 @@ module Locabulary
 
     def convert_to_json(data)
       json_array = data.map do |row|
-        Locabulary::Item.new(row).to_h
+        Locabulary::Items.build(row).to_h
       end
       @json_data = JSON.pretty_generate("predicate_name" => predicate_name, "values" => json_array)
     end
