+++ spec/lib/locabulary/items_spec.rb
@@ -2,6 +2,17 @@ require 'spec_helper'
 require 'locabulary/items'
 
 RSpec.describe Locabulary::Items do
+  context '#build' do
+    it "finds a defined subclass" do
+      expect(described_class.build(predicate_name: 'administrative_units')).to(
+        be_a(Locabulary::Items::AdministrativeUnit)
+      )
+    end
+
+    it "defaults to the Items::Base" do
+      expect(described_class.build(predicate_name: 'chicken')).to(be_a(Locabulary::Items::Base))
+    end
+  end
   context '#builder_for' do
     it "finds a defined subclass" do
       expect(described_class.builder_for(predicate_name: 'administrative_units').call).to(
@@ -9,8 +20,8 @@ RSpec.describe Locabulary::Items do
       )
     end
 
-    it "defaults to the base Locabulary::Item" do
-      expect(described_class.builder_for(predicate_name: 'chicken').call).to(be_a(Locabulary::Item))
+    it "defaults to the Items::Base" do
+      expect(described_class.builder_for(predicate_name: 'chicken').call).to(be_a(Locabulary::Items::Base))
     end
   end
 end
